scan_sign: get_byte instead of get_int
scan_proc, scan_label, scan_var: size_t/ptrdiff_t instead of U/WIDE/INT64, with range check
  I would like these to all be size_t but I think we use negative values
  in some places, that are to be ignored or somesuch.

rejigger types a bit:
 mostly WIDE => INT64, UWIDE => UINT64
 but then some casts to WIDE/UWIDE for printf
   (I considered being independent of that by using our own
   formating functions and string concat (ACONCAT) but nevermind, tedious
   and should get on to more valuable stuff.)

fix warning in m3_do_fixed_insert

ULONG => UINT/UINT32
  use limits.h to find UINT32 (ie: actually still portable to 16 bit int)

+++ m3-sys/m3cc/gcc/gcc/m3cg/m3gty45.h
@@ -65,6 +65,6 @@ struct GTY(()) language_function
 
 typedef struct GTY(()) m3type
 {
-  unsigned long type_id;
+  unsigned type_id;
   tree GTY(()) t;
 } m3type_t;
